+++ components.d.ts
@@ -10,6 +10,7 @@ declare module 'vue' {
     AppHeader: typeof import('./src/components/AppHeader.vue')['default']
     PokemonCard: typeof import('./src/components/PokemonCard.vue')['default']
     PokemonSearch: typeof import('./src/components/PokemonSearch.vue')['default']
+    PokemonStats: typeof import('./src/components/PokemonStats.vue')['default']
     PokemonTypesChips: typeof import('./src/components/PokemonTypesChips.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
